refactor(ui): canonicalize totp manage and tab module filenames
